Extract base URL logic into new abstract UrlImageEditor
