+++ image/src/main/java/com/psddev/dari/util/JavaImageEditor.java
@@ -24,7 +24,7 @@ import org.imgscalr.Scalr;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class JavaImageEditor extends AbstractImageEditor {
+public class JavaImageEditor extends AbstractUrlImageEditor {
 
     private static final String DEFAULT_IMAGE_FORMAT = "png";
     private static final String DEFAULT_IMAGE_CONTENT_TYPE = "image/" + DEFAULT_IMAGE_FORMAT;
@@ -41,7 +41,6 @@ public class JavaImageEditor extends AbstractImageEditor {
     protected static final String THUMBNAIL_COMMAND = "thumbnail";
 
     private Scalr.Method quality = Scalr.Method.AUTOMATIC;
-    private String baseUrl;
     private String basePath;
     private String sharedSecret;
     private String errorImage;
@@ -57,10 +56,6 @@ public class JavaImageEditor extends AbstractImageEditor {
         this.quality = quality;
     }
 
-    public String getBaseUrl() {
-        return baseUrl;
-    }
-
     public String getBasePath() {
         if (StringUtils.isBlank(basePath) && !(StringUtils.isBlank(baseUrl))) {
             basePath = baseUrl.substring(baseUrl.indexOf("//") + 2);
@@ -70,10 +65,6 @@ public class JavaImageEditor extends AbstractImageEditor {
             }
         }
         return basePath;
-    }
-
-    public void setBaseUrl(String baseUrl) {
-        this.baseUrl = baseUrl;
     }
 
     public String getSharedSecret() {
@@ -350,10 +341,6 @@ public class JavaImageEditor extends AbstractImageEditor {
             } else if (qualitySetting instanceof String) {
                 quality = Scalr.Method.valueOf(ObjectUtils.to(String.class, qualitySetting));
             }
-        }
-
-        if (!ObjectUtils.isBlank(settings.get("baseUrl"))) {
-            setBaseUrl(ObjectUtils.to(String.class, settings.get("baseUrl")));
         }
 
         if (!ObjectUtils.isBlank(settings.get("sharedSecret"))) {
